[FEAT]: finish tests for PollService

+++ Backend/src/main/java/com/m347/pollit/services/PollService.java
@@ -105,7 +105,7 @@ public class PollService {
                 case SLIDER:
                 case NUMBER:
                     try {
-                        Long.parseLong(answer);
+                        Float.parseFloat(answer);
                         addAnswer(element, answer);
                     } catch (NumberFormatException e) {
                         throw new CommonException("Bitte eine gültige Zahl eingeben");
